- Added cmake support - Tested with clang 3.4 and gccc 4.8 (under Ubuntu 13.10), x86 and amd64 - Fix a few misc bugs in lzhamtest (wrong deinit funcs called during compression when the src file was unreadable) - Adding src file readability check to lzhamtest - lzhamtest now defaults to sending down -1 as the max # of helper threads, so the lib should use all available cores

+++ lzhamdecomp/lzham_vector.cpp
@@ -21,8 +21,8 @@ namespace lzham
          return true;
 
       size_t new_capacity = min_new_capacity;
-      if ((grow_hint) && (!math::is_power_of_2(new_capacity)))
-         new_capacity = math::next_pow2(new_capacity);
+      if ((grow_hint) && (!math::is_power_of_2(static_cast<uint64>(new_capacity))))
+         new_capacity = math::next_pow2(static_cast<uint64>(new_capacity));
 
       LZHAM_ASSERT(new_capacity && (new_capacity > m_capacity));
